refactor: breakdown the god class  into the components
